backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Pillow==11.1.0
gunicorn==23.0.0
django-storages[boto3]==1.14.4

